Moved option picks metadata to storage
